mocking test step 4: amend tests based on PR comments
