+++ infra/docker/core/Dockerfile.dev
@@ -1,6 +1,6 @@
 FROM openjdk:11-jre
 ARG REVISION=dev
-ADD $PWD/core/target/feast-core-$REVISION-exec.jar /opt/feast/feast-core.jar
+ADD $PWD/core/target/feast-core-$REVISION.jar /opt/feast/feast-core.jar
 CMD ["java",\
      "-Xms2048m",\
      "-Xmx2048m",\
