+++ infra/docker/serving/Dockerfile.dev
@@ -7,7 +7,7 @@ ARG REVISION=dev
 RUN wget -q https://github.com/grpc-ecosystem/grpc-health-probe/releases/download/v0.3.1/grpc_health_probe-linux-amd64 \
          -O /usr/bin/grpc-health-probe && \
     chmod +x /usr/bin/grpc-health-probe
-ADD $PWD/serving/target/feast-serving-$REVISION-exec.jar /opt/feast/feast-serving.jar
+ADD $PWD/serving/target/feast-serving-$REVISION.jar /opt/feast/feast-serving.jar
 CMD ["java",\
      "-Xms1024m",\
      "-Xmx1024m",\
