+++ serving/pom.xml
@@ -253,6 +253,7 @@
       <groupId>com.amazonaws</groupId>
       <artifactId>aws-java-sdk</artifactId>
       <version>1.11.827</version>
+      <scope>test</scope>
     </dependency>
 
     <dependency>
@@ -350,6 +351,7 @@
     <dependency>
        <groupId>org.springframework.kafka</groupId>
        <artifactId>spring-kafka</artifactId>
+       <scope>test</scope>
     </dependency>
   </dependencies>
