+++ serving/src/test/resources/application-it.properties
@@ -16,3 +16,23 @@
 feast.core-authentication.enabled=false
 feast.security.authentication.enabled=false
 feast.security.authorization.enabled=false
+
+# For ServingServiceSnowflakeStorageIT
+feast.active_store=historical_snowflake
+feast.stores=[{\
+  name=historical_snowflake, \
+  type=JDBC, \
+  config={\
+    schema=PUBLIC, \
+    database=DEMO_DB, \
+    role=ACCOUNTADMIN, \
+    table_name=FEAST_FEATURES_TEST, \
+    url=jdbc:snowflake://nx46274.us-east-2.aws.snowflakecomputing.com, \
+    username=SNOWFLAKETEST, \
+    password=123456Pw, \
+    warehouse=COMPUTE_WH, \
+    staging_location=s3://feast-snowflake-staging/test/, \
+    class_name=net.snowflake.client.jdbc.SnowflakeDriver, \
+    batch_size=1, \
+    storage_integration=s3_int}, \
+  subscriptions=[{name=*, project=*}]}]
