--- serving/src/test/resources/application-it.yml
@@ -1,26 +1,0 @@
-feast:
-
-  # Indicates the active store. Only a single store in the last can be active at one time. In the future this key
-  # will be deprecated in order to allow multiple stores to be served from a single serving instance
-  active_store: historical_snowflake
-
-  # List of store configurations
-  stores:
-    - name: historical_snowflake
-      type: JDBC
-      config:  # Store specific configuration.
-        schema: PUBLIC
-        database: DEMO_DB
-        role: ACCOUNTADMIN
-        table_name: FEAST_FEATURES_TEST
-        url: jdbc:snowflake://nx46274.us-east-2.aws.snowflakecomputing.com
-        username: SNOWFLAKETEST
-        password: 123456Pw
-        warehouse: COMPUTE_WH
-        staging_location: s3://feast-snowflake-staging/test/
-        class_name: net.snowflake.client.jdbc.SnowflakeDriver
-        batch_size: 1
-        storage_integration: s3_int
-      subscriptions:
-        - name: "*"
-          project: "*"
