+++ serving/src/test/resources/docker-compose/docker-compose-it-core.yml
@@ -2,7 +2,7 @@ version: '3'
 
 services:
   core:
-    image: feast-core:dev1
+    image: gcr.io/kf-feast/feast-core:latest
     volumes:
       - ./core/application-it.yml:/etc/feast/application.yml
     environment:
