+++ storage/connectors/jdbc/pom.xml
@@ -136,6 +136,13 @@
             <artifactId>mockito-core</artifactId>
             <version>2.8.9</version>
         </dependency>
+        <!--  aws s3 client for testing  -->
+        <dependency>
+            <groupId>com.amazonaws</groupId>
+            <artifactId>aws-java-sdk</artifactId>
+            <version>1.11.827</version>
+            <scope>test</scope>
+        </dependency>
   		<dependency>
 	        <groupId>org.json</groupId>
 	        <artifactId>json</artifactId>
@@ -147,7 +154,4 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-    <properties>
-        <maven.test.skip>true</maven.test.skip>
-    </properties>
 </project>
