+++ storage/connectors/jdbc/src/test/java/feast/storage/connectors/jdbc/retriever/SnowflakeHistoricalRetrieverTest.java
@@ -23,27 +23,33 @@ import feast.proto.core.FeatureSetProto;
 import feast.proto.serving.ServingAPIProto;
 import feast.storage.api.retriever.FeatureSetRequest;
 import feast.storage.api.retriever.HistoricalRetrievalResult;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
+import java.io.IOException;
+import java.util.*;
+import java.util.zip.GZIPInputStream;
+import net.snowflake.client.jdbc.internal.amazonaws.auth.profile.ProfileCredentialsProvider;
+import net.snowflake.client.jdbc.internal.amazonaws.services.s3.AmazonS3;
+import net.snowflake.client.jdbc.internal.amazonaws.services.s3.AmazonS3ClientBuilder;
+import net.snowflake.client.jdbc.internal.amazonaws.services.s3.AmazonS3URI;
+import net.snowflake.client.jdbc.internal.amazonaws.services.s3.model.GetObjectRequest;
+import net.snowflake.client.jdbc.internal.amazonaws.services.s3.model.S3Object;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-public class SnowflakeHistoricalRetrieverJSONColTest {
+@Ignore
+public class SnowflakeHistoricalRetrieverTest {
+  /** Manual tests needs a testing Snowflake account */
   private JdbcHistoricalRetriever snowflakeFeatureRetriever;
   //  Snowflake account
-  //  private String staging_location = System.getenv("STAGING_LOCATION");
   private String staging_location = "s3://feast-snowflake-staging/test/";
 
   private Map<String, String> snowflakeConfig = new HashMap<>();
-  private String SFUrl = "jdbc:snowflake://nx46274.us-east-2.aws.snowflakecomputing.com";
+  private String SFUrl = System.getenv("SNOWFLAKE_URL");
   private String SFClassName = "net.snowflake.client.jdbc.SnowflakeDriver";
-  private String SFusername = System.getenv("SNOWFLAKE_USERNAME_RETRI");
-  private String SFpw = System.getenv("SNOWFLAKE_PASSWORD_RETRI");
+  private String SFUsername = System.getenv("SNOWFLAKE_USERNAME");
+  private String SFpw = System.getenv("SNOWFLAKE_PASSWORD");
   private String SFDatabase = "DEMO_DB";
   private String SFSchema = "PUBLIC";
   private String SFTable = "FEAST_FEATURES";
@@ -56,7 +62,7 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
     snowflakeConfig.put("database", SFDatabase);
     snowflakeConfig.put("schema", SFSchema);
     snowflakeConfig.put("class_name", SFClassName);
-    snowflakeConfig.put("username", SFusername);
+    snowflakeConfig.put("username", SFUsername);
     snowflakeConfig.put("password", SFpw);
     snowflakeConfig.put("url", SFUrl);
     snowflakeConfig.put("staging_location", staging_location);
@@ -84,7 +90,7 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
   }
 
   @Test
-  public void shouldRetrieveFromSnowflakeTest3DatesWithMaxAge() {
+  public void shouldRetrieveFromSnowflakeTest3DatesWithMaxAge() throws IOException {
     String entitySourceUri = "s3://feast-snowflake-staging/test/entity_tables/entities_3dates.csv";
     ServingAPIProto.DatasetSource.FileSource fileSource =
         ServingAPIProto.DatasetSource.FileSource.newBuilder()
@@ -103,10 +109,15 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
     List<String> files = snowflakeHisRetrievalResult.getFileUris();
     /** Should return ENTITY_ID_PRIMARY, FEATURE_SET__FEATURE_1 1,null 2,100 3,300 */
     Assert.assertTrue(files.get(0).contains(staging_location));
+    // get csv.gz file from s3
+    List<String> resultLines = this.readFromS3(files.get(0));
+    Assert.assertEquals("\\" + "\\N", resultLines.get(1).split(",")[3]);
+    Assert.assertEquals("100", resultLines.get(2).split(",")[3]);
+    Assert.assertEquals("300", resultLines.get(3).split(",")[3]);
   }
 
   @Test
-  public void shouldRetrieveFromSnowflakeTest1Date() {
+  public void shouldRetrieveFromSnowflakeTest1Date() throws IOException {
     //      Set CSV format DATA_FORMAT_CSV = 2; where the first column of the csv file must be
     // entity_id
     //      file_uri is under
@@ -130,10 +141,15 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
     List<String> files = snowflakeHisRetrievalResult.getFileUris();
     /** Should return ENTITY_ID_PRIMARY, FEATURE_SET__FEATURE_1 1,410 2,220 3,300 */
     Assert.assertTrue(files.get(0).contains(staging_location));
+    // get csv.gz file from s3
+    List<String> resultLines = this.readFromS3(files.get(0));
+    Assert.assertEquals("410", resultLines.get(1).split(",")[3]);
+    Assert.assertEquals("220", resultLines.get(2).split(",")[3]);
+    Assert.assertEquals("300", resultLines.get(3).split(",")[3]);
   }
 
   @Test
-  public void shouldRetrieveFromSnowflakeTest1DateWithNull() {
+  public void shouldRetrieveFromSnowflakeTest1DateWithNull() throws IOException {
     //      Set CSV format DATA_FORMAT_CSV = 2; where the first column of the csv file must be
     // entity_id
     //      file_uri is under
@@ -158,10 +174,15 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
     List<String> files = snowflakeHisRetrievalResult.getFileUris();
     /** Should return ENTITY_ID_PRIMARY, FEATURE_SET__FEATURE_1 1,410 2,100 3,null */
     Assert.assertTrue(files.get(0).contains(staging_location));
+    // get csv.gz file from s3
+    List<String> resultLines = this.readFromS3(files.get(0));
+    Assert.assertEquals("410", resultLines.get(1).split(",")[3]);
+    Assert.assertEquals("100", resultLines.get(2).split(",")[3]);
+    Assert.assertEquals("\\" + "\\N", resultLines.get(3).split(",")[3]);
   }
 
   @Test
-  public void shouldRetrieveFromSnowflakeTestSameIdsWithMaxAge() {
+  public void shouldRetrieveFromSnowflakeTestSameIdsWithMaxAge() throws IOException {
     //      Set CSV format DATA_FORMAT_CSV = 2; where the first column of the csv file must be
     // entity_id
     //      file_uri is under
@@ -184,6 +205,12 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
     List<String> files = snowflakeHisRetrievalResult.getFileUris();
     /** Should return ENTITY_ID_PRIMARY, FEATURE_SET__FEATURE_1 1,null 2,100 3,300 1,410 */
     Assert.assertTrue(files.get(0).contains(staging_location));
+    // get csv.gz file from s3
+    List<String> resultLines = this.readFromS3(files.get(0));
+    Assert.assertEquals("\\" + "\\N", resultLines.get(1).split(",")[3]);
+    Assert.assertEquals("100", resultLines.get(2).split(",")[3]);
+    Assert.assertEquals("300", resultLines.get(3).split(",")[3]);
+    Assert.assertEquals("410", resultLines.get(4).split(",")[3]);
   }
 
   private List<FeatureSetRequest> createFeatureSetRequests() {
@@ -249,5 +276,28 @@ public class SnowflakeHistoricalRetrieverJSONColTest {
         .addFeatures(FeatureSetProto.FeatureSpec.newBuilder().setName("feature_1"))
         .addFeatures(FeatureSetProto.FeatureSpec.newBuilder().setName("feature_2"))
         .build();
+  }
+
+  /**
+   * get the result feature set in s3 as a list of lines
+   *
+   * @param fileUri
+   * @return
+   * @throws IOException
+   */
+  private List<String> readFromS3(String fileUri) throws IOException {
+    AmazonS3 s3client =
+        AmazonS3ClientBuilder.standard().withCredentials(new ProfileCredentialsProvider()).build();
+    String bucket = new AmazonS3URI(fileUri).getBucket();
+    String key = new AmazonS3URI(fileUri).getKey();
+    S3Object fileObj = s3client.getObject(new GetObjectRequest(bucket, key));
+    Scanner fileIn = new Scanner(new GZIPInputStream(fileObj.getObjectContent()));
+    List<String> resultLines = new ArrayList<>();
+    if (null != fileIn) {
+      while (fileIn.hasNext()) {
+        resultLines.add(fileIn.nextLine());
+      }
+    }
+    return resultLines;
   }
 }
