--- storage/connectors/jdbc/src/test/java/feast/storage/connectors/jdbc/retriever/myproject2_entity_rows.csv
@@ -1,4 +1,0 @@
-entity_id_primary,created_timestamp
-4,2020-06-10 15:52:33.009000
-4,2020-06-10 16:06:03.808000
-4,2020-06-10 17:58:29.343000
